requirements.txt
streamlit>=1.35.0
fastapi>=0.110.0
uvicorn>=0.30.0
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
plotly>=5.20.0
pymongo>=4.7.0
python-dotenv>=1.0.1
pydantic>=2.7.0
google-generativeai>=0.8.0
pytest>=8.0.0
folium>=0.14.0
streamlit-folium>=0.11.0
apscheduler>=3.10.4
supabase